🚚 refactor: move self cli setup out of agent
