+++ internal/cli/selfcli.go
@@ -1,4 +1,4 @@
-package agent
+package cli
 
 import (
 	"fmt"
