+++ internal/cli/selfcli_test.go
@@ -1,4 +1,4 @@
-package agent
+package cli
 
 import (
 	"os"
